Rename init command directory to match package name
Minor fixup since `init` is not a valid package name (reserved keyword
in Go)
